修复B站解析问题

+++ src/main/kotlin/plugin/command/card/BiliBili.kt
@@ -283,7 +283,7 @@ data class News(
     val source_url: String,
     val tag: String,
     val title: String,
-    val uin: Int
+    val uin: Long
 )
 
 data class BiliSearchApi(
